remove play method from TTT class
