+++ lib/tic_tac_toe.rb
@@ -1,15 +1,14 @@
 require_relative './command_line_in'
 require_relative './prompt'
-require_relative './board_presenter'
 require_relative './board'
 require_relative './input_validator'
 require_relative './player'
+require_relative './game'
 
 class TicTacToe
     attr_accessor :board 
     def initialize
         @board = Board.new
-        @presenter = BoardPresenter.new
         @players = []
     end
 
@@ -26,21 +25,6 @@ class TicTacToe
         end
     end
 
-    def select_coordinate
-        puts Prompt::MAKE_COORDINATE_SELECTION
-        CommandLineIn.new.get_input.to_i
-    end
-
-    def take_turn(board, position, token)
-        input_validator = InputValidator.new
-        if input_validator.valid_coordinate?(@board.moves, position)
-            @board.move(token, position)
-        else
-            user_coordinate = select_coordinate
-            take_turn(@board.moves, user_coordinate, token)
-        end
-    end
-
     def start_game_engine
         puts Prompt::WELCOME
         2.times { |i|
@@ -49,17 +33,7 @@ class TicTacToe
             selected_token = CommandLineIn.new.get_input.upcase
             assign_player_token(selected_token)
         }
-        play_game
-    end
-
-    def play_game
-        @players.each { |player| 
-            puts "HELLO PLAYER #{player.token}"
-            puts @presenter.display_board(@board.moves)
-            user_coordinate = select_coordinate
-            take_turn(@board.moves, user_coordinate, player.token)
-        }
-        play_game
+        Game.new.play(@players, @board)
     end
 end
